Convert cost function to Double
otherwise we get surprising floating point rounding error

+++ src/ksc/Ksc/Cost.hs
@@ -5,7 +5,7 @@ module Ksc.Cost where
 import qualified Lang
 import           Prim (isThePrimFun)
 
-cost :: Lang.TExpr -> Either String Float
+cost :: Lang.TExpr -> Either String Double
 cost = \case
   Lang.Konst{}  -> pure 0
   Lang.Var{}    -> pure 0
@@ -45,7 +45,7 @@ size = \case
   Lang.TypeUnknown -> notImpl "Unknown"
   where notImpl s = Left ("size not implemented for " ++ s)
 
-costFun :: Lang.TFun -> Lang.Type -> Either String Float
+costFun :: Lang.TFun -> Lang.Type -> Either String Double
 costFun f t
   | f `isThePrimFun` "add"    = fmap fromIntegral (size t)
   | f `isThePrimFun` "and"    = fmap fromIntegral (size t)
